+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -8,6 +8,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Added
 
+- Added the new "All Inclusive" bundle to Extensions (#3269)
 - JetEngine extension for Gato GraphQL (#3264)
 - Support executing bulk actions on users page (#3261)